CrownShop/requirements.txt
appdirs==1.4.3
asgiref==3.2.10
CacheControl==0.12.6
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.1.1
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.2
django-allauth==0.43.0
django-cors-headers==3.5.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-debug-toolbar==3.1.1
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
Pillow==8.0.1
progress==1.5
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
python-decouple==3.3
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
urllib3==1.25.11
webencodings==0.5.1
